refactor crud queries into separate folder
